requirements.txt
flask==3.0.0
gunicorn==21.2.0
pyotp==2.9.0
bcrypt==4.1.2
markdown2==2.4.12
werkzeug==3.0.1
qrcode==8.2
pillow==12.0.0
python-dotenv==1.2.1